# api/swagger.yaml
swagger: "2.0"
info:
  description: "Simple REST API for recepies"
  version: "0.0.1"
  title: "Recepies API"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
basePath: "/"
tags:
- name: "recipe"
  description: "Recipe"
schemes:
- "http"
paths:
  /recipe:
    post:
      tags:
      - "recipe"
      summary: "Create recipe"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "recipe"
        in: "body"
        description: "Recipe object to create"
        required: true
        schema:
          $ref: "#/definitions/Recipe"
      responses:
        200:
          description: "Successful operation"
          schema:
            $ref: "#/definitions/Recipe"
        405:
          description: "Invalid input"
  /recipe/{id}:
    get:
      tags:
      - "recipe"
      summary: "Read recipe"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "ID of recipe"
        required: true
        type: string
      responses:
        200:
          description: "Successful operation"
          schema:
            $ref: "#/definitions/Recipe"
        400:
          description: "Invalid ID"
        404:
          description: "Recipe not found"
    put:
      tags:
      - "recipe"
      summary: "Update recipe"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "ID of recipe"
        required: true
        type: string
      - name: "recipe"
        in: "body"
        description: "Recipe object to create"
        required: true
        schema:
          $ref: "#/definitions/Recipe"
      responses:
        200:
          description: "Successful operation"
          schema:
            $ref: "#/definitions/Recipe"
        400:
          description: "Invalid ID"
        404:
          description: "Recipe not found"
        405:
          description: "Invalid input"
    delete:
      tags:
      - "recipe"
      summary: "Delete recipe"
      parameters:
      - name: "id"
        in: "path"
        description: "ID of recipe"
        required: true
        type: string
      responses:
        400:
          description: "Invalid ID"
        404:
          description: "Recipe not found"
  /recipe/{id}/upload:
    post:
      tags:
      - "recipe"
      summary: "Upload image"
      consumes:
         - multipart/form-data
      parameters:
        - name: "id"
          in: "path"
          description: "ID of recipe"
          required: true
          type: string
        - in: formData
          name: image
          type: file
          description: The file to upload.
      responses:
        400:
          description: "Invalid ID"
        404:
          description: "Recipe not found"
        415:
          description: "Illegal content type"
  /recepies:
    get:
      tags:
      - "recepies"
      summary: "List recipes"
      produces:
      - "application/json"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Recipe"
  /recepies/names:
    get:
      tags:
      - "recepies"
      summary: "List recepie names"
      produces:
      - "application/json"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/RecipeNames"
  /ingredients:
    get:
      tags:
      - "ingredients"
      summary: "List all distinct ingredients for all recepies"
      produces:
      - "application/json"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              type: "string"
definitions:
  Recipe:
    properties:
      id:
        type: "string"
        description: "BSON ObjectID"
      title:
        type: "string"
        description: "The recipe title"
      description:
        type: "string"
        description: "The recipe description"
      url:
        type: "string"
        description: "URL to the original recipe"
      ingredient:
        type: "array"
        items:
          $ref: "#/definitions/Ingredient"
  RecipeNames:
    properties:
      id:
        type: "string"
        description: "BSON ObjectID"
      title:
        type: "string"
        description: "The recipe title"
  Ingredient:
    properties:
      amount:
        type: "integer"
        format: "uint16"
        description: "Amount of ingredient"
      unit:
        type: "string"
        description: "Unit of the ingredient"
      name:
        type: "string"
        description: "Name of the ingredient"